break cloning registry into its own file
