+++ src/logger.js
@@ -1,0 +1,5 @@
+var logger = require('davlog')
+
+logger.init({name: 'registry-mirror'})
+
+module.exports = logger
